Move parser to common folder
