+++ src/validator.ts
@@ -7,7 +7,7 @@
 import path from "path";
 import amf from "amf-client-js";
 
-import { parseRamlFile } from "./parser";
+import { parseRamlFile } from "./common/parser";
 import { profilePath } from ".";
 
 export async function validateCustom(
